change variable names
This is update that Matma coded for us based on the variable names we agreed on.  The only difference from the meeting was that "8grade" and "3grade" had to be changed to "grade3" and "grade8".  This is because default variable names in R apparently should shart with letters instead of numbers.  Else you would need to do `variablename` which can be annoying to remember.