lpsolve cleanup for Eric and Rshiny
TLDR:  final replaced "linear" and "test2" in the previous code so we read all the same code.  lptest replaced "test".
As a result of the talking to the people doing Rshiny, I am cleaning up the lp solve so that the lp solver testfuction is called lptest instead of test .  In turn, the final results will be called final instead of test2 or linear.

This way it make more sense to the people doing Rshiny.
in the same way eventually df0 and df2 will be replaced.
#The only reason that df2 not become original_constrants is because (orginal constrants is a direct transformation of orignal data! (df0 --> df2)